Revert "[Tizen] Fix Svace issue for 64bit"
This reverts commit dfb16c0dbc52b529ac39869a667524605f5ea0bb.

+++ dali/public-api/rendering/texture-set.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2020 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -86,7 +86,7 @@ Sampler TextureSet::GetSampler(size_t index) const
   return Dali::Sampler(samplerPtr);
 }
 
-uint32_t TextureSet::GetTextureCount() const
+size_t TextureSet::GetTextureCount() const
 {
   return GetImplementation(*this).GetTextureCount();
 }
